requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1.7
djangorestframework==3.10.3
idna==2.8
pika==1.1.0
pytz==2018.9
requests==2.22.0
urllib3==1.25.6
